requirements.txt
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
dnspython==2.4.2
exceptiongroup==1.1.3
filelock==3.12.4
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
outcome==1.3.0
packaging==23.2
parsel==1.8.1
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pymongo==4.5.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.0
scrapy-selenium==0.0.7
scrapyd==1.4.3
scrapyd-client==1.2.3
selenium==4.14.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tldextract==3.6.0
trio==0.22.2
trio-websocket==0.11.1
Twisted==22.10.0
typing_extensions==4.8.0
uberegg==0.1.1
urllib3==2.0.6
w3lib==2.1.2
webdriver-manager==4.0.1
wsproto==1.2.0
zope.interface==6.1
